chore: move interface
