+++ app/application/usecase/card/card.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/RyoheiTomiyama/phraze-api/domain"
-	"github.com/RyoheiTomiyama/phraze-api/infra/db"
+	"github.com/RyoheiTomiyama/phraze-api/domain/infra/db"
 	"github.com/RyoheiTomiyama/phraze-api/infra/gemini"
 	"github.com/RyoheiTomiyama/phraze-api/service/card"
 )
