+++ app/application/usecase/deck/deck.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/RyoheiTomiyama/phraze-api/domain"
-	"github.com/RyoheiTomiyama/phraze-api/infra/db"
+	"github.com/RyoheiTomiyama/phraze-api/domain/infra/db"
 	"github.com/RyoheiTomiyama/phraze-api/util/auth"
 	"github.com/RyoheiTomiyama/phraze-api/util/errutil"
 )
