+++ app/infra/db/client.go
@@ -1,13 +1,11 @@
 package db
 
 import (
-	"context"
 	"database/sql"
 	"fmt"
 	"testing"
-	"time"
 
-	"github.com/RyoheiTomiyama/phraze-api/domain"
+	"github.com/RyoheiTomiyama/phraze-api/domain/infra/db"
 	_ "github.com/jackc/pgx/v5/stdlib"
 	"github.com/jmoiron/sqlx"
 )
@@ -16,38 +14,7 @@ type client struct {
 	db *sqlx.DB
 }
 
-type IClient interface {
-	/* cards */
-	CreateCard(ctx context.Context, card *domain.Card) (*domain.Card, error)
-	GetCard(ctx context.Context, id int64) (*domain.Card, error)
-	GetCards(ctx context.Context, where *domain.CardsWhere, limit, offset int) ([]*domain.Card, error)
-	GetPendingCards(ctx context.Context, deckID int64, to time.Time, limit, offset int) ([]*domain.Card, error)
-	UpdateCardByID(ctx context.Context, id int64, input *domain.UpdateCardInput) (*domain.Card, error)
-	CountCards(ctx context.Context, where *domain.CardsWhere) (int, error)
-	DeleteCard(ctx context.Context, id int64) (int64, error)
-
-	/* card_reviews */
-	CreateCardReview(ctx context.Context, review *domain.CardReview) (*domain.CardReview, error)
-	GetLatestCardReview(ctx context.Context, cardID int64) (*domain.CardReview, error)
-
-	/* card_schedules */
-	GetCardSchedule(ctx context.Context, cardID int64) (*domain.CardSchedule, error)
-	GetLatestCardSchedulesByDeckID(ctx context.Context, deckIDs []int64) (map[int64]*domain.CardSchedule, error)
-	UpsertCardSchedule(ctx context.Context, schedule *domain.CardSchedule) (*domain.CardSchedule, error)
-
-	CreateDeck(ctx context.Context, deck *domain.Deck) (*domain.Deck, error)
-	GetDeck(ctx context.Context, id int64) (*domain.Deck, error)
-	GetDecks(ctx context.Context, userID string) ([]*domain.Deck, error)
-	DeleteDeck(ctx context.Context, id int64) (int64, error)
-
-	GetDeckInfosByDeckID(ctx context.Context, deckIDs []int64) (map[int64]*domain.DeckInfo, error)
-
-	/* permissions */
-	GetPermissionsByUserID(ctx context.Context, userID string) ([]*domain.Permission, error)
-	HasPermissionByUserID(ctx context.Context, userID string, key domain.PermissionKey) (bool, error)
-
-	Tx(ctx context.Context, fn func(ctx context.Context) error) error
-}
+var _ db.IClient = (*client)(nil)
 
 type DataSourceOption struct {
 	Host     string
@@ -57,7 +24,7 @@ type DataSourceOption struct {
 	Password string
 }
 
-func NewClient(opt DataSourceOption) (IClient, error) {
+func NewClient(opt DataSourceOption) (db.IClient, error) {
 	db, err := open(opt)
 	if err != nil {
 		return nil, err
@@ -85,6 +52,6 @@ func open(opt DataSourceOption) (*sqlx.DB, error) {
 	return db, nil
 }
 
-func NewTestClient(t *testing.T, db *sqlx.DB) IClient {
+func NewTestClient(t *testing.T, db *sqlx.DB) db.IClient {
 	return &client{db}
 }
